requirements.txt
# ===========================
# Core
# ===========================
pytest==8.1.1
playwright==1.44.0
pytest-playwright==0.5.0

# ===========================
# Allure Reporting
# ===========================
allure-pytest==2.13.5

# ===========================
# API Testing
# ===========================
requests==2.31.0
jsonschema==4.22.0

# ===========================
# Config
# ===========================
PyYAML==6.0.1
python-dotenv==1.0.1

# ===========================
# Mobile (Appium) — optional
# Install separately if needed:
# pip install appium-python-client==3.1.0
# pip install selenium==4.21.0
# ===========================
# appium-python-client==3.1.0
# selenium==4.21.0

# ===========================
# Parallel execution
# ===========================
pytest-xdist==3.5.0

# ===========================
# Retry flaky tests
# ===========================
pytest-rerunfailures==14.0

# ===========================
# HTML backup report
# ===========================
pytest-html==4.1.1
